Create a node list specification

Creates a node list specification that is used by the provided estimators. create_node_list() is not explicitly called by the analyst, rather it is provided so the analyst can confirm how estimators will use variables before actually performing the estimation procedure.

create_node_list(trt, nodes, baseline = NULL, k = Inf)

Arguments

trt
A vector of column names of treatment variables.

nodes
A list of length tau with the column names for new nodes to be introduced at each time point. The list should be ordered following the time ordering of the model.

baseline
An optional vector of columns names for baseline covariates to be included for adjustment at every timepoint.

k
An integer specifying how many previous time points should be used for estimation at the given time point. Default is Inf, all time points.

Value

A list the same length of the nodes parameter with the variables to be used for estimation at that given time point.

Examples

a <- c("A_1", "A_2", "A_3", "A_4")
bs <- c("W_1", "W_2")
nodes <- list(c("L_1"), c("L_2"), c("L_3"), c("L_4"))

# assuming no Markov property
create_node_list(a, nodes, bs, k = Inf)

#> [[1]]
#> [1] "W_1" "W_2" "L_1" "A_1"
#> 
#> [[2]]
#> [1] "W_1" "W_2" "L_1" "A_1" "L_2" "A_2"
#> 
#> [[3]]
#> [1] "W_1" "W_2" "L_1" "A_1" "L_2" "A_2" "L_3" "A_3"
#> 
#> [[4]]
#>  [1] "W_1" "W_2" "L_1" "A_1" "L_2" "A_2" "L_3" "A_3" "L_4" "A_4"
#> 

# assuming a Markov property
create_node_list(a, nodes, bs, k = 1)

#> [[1]]
#> [1] "W_1" "W_2" "L_1" "A_1"
#> 
#> [[2]]
#> [1] "W_1" "W_2" "L_1" "A_1" "L_2" "A_2"
#> 
#> [[3]]
#> [1] "W_1" "W_2" "L_2" "A_2" "L_3" "A_3"
#> 
#> [[4]]
#> [1] "W_1" "W_2" "L_3" "A_3" "L_4" "A_4"
#>