lmtp: Non-parametric Causal Effects of Feasible Interventions Based on Modified Treatment Policies

Non-parametric estimators for casual effects based on longitudinal modified treatment policies as described in Diaz, Williams, and Hoffman (2020) <arXiv:https://arxiv.org/abs/2006.01366>, traditional point treatment, and traditional longitudinal effects. Continuous, binary, and categorical treatments are allowed as well are censored outcomes. The treatment mechanism is estimated via a density ratio classification procedure irrespective of treatment variable type. Estimation is performed using the Super Learner from `sl3`. For both continuous and binary outcomes, additive treatment effects can be calculated and relative risks and odds ratios may be calculated for binary outcomes.

Author

Maintainer: Nicholas Williams niw4001@med.cornell.edu (ORCID)

Authors:

Iván Díaz ild2005@med.cornell.edu (ORCID)