Perform Contrasts of LMTP Fits

Estimates contrasts of multiple LMTP fits compared to either a known reference value or a reference LMTP fit.

lmtp_contrast(..., ref, type = c("additive", "rr", "or"))

Arguments

...
One or more objects of class lmtp.

ref
A reference value or another lmtp fit to compare all other fits against.

type
The contrasts of interest. Options are "additive" (the default), "rr", and "or".

Value

A list of class lmtp_contrast containing the following components:

type

The type of contrast performed.

null

The null hypothesis.

vals

A dataframe containing the contrasts estimates, standard errors, and confidence intervals.

eifs

Un-centered estimated influence functions for contrasts estimated.

Examples

# \donttest{
library(lmtp)
head(sim_cens)

#>   L1         A1 C1 L2         A2 C2 Y
#> 1  0  0.7930531  1  0  0.4098544  1 0
#> 2  1  3.9568969  1  1  7.1306144  1 1
#> 3  0 -0.9803504  1  0 -2.7858973  1 0
#> 4  1  0.8833268  1  1  1.1372421  1 1
#> 5  1  3.1666751  1  1  7.4640462  1 1
#> 6  1  3.5393771  1  1  9.5161813  1 1

a <- c("A1", "A2")
nodes <- list(c("L1"), c("L2"))
cens <- c("C1", "C2")
y <- "Y"
# mean population outcome
psi_null <- lmtp_tmle(sim_cens, a, y, time_vary = nodes,
                      cens = cens, shift = NULL, folds = 2)

# treatment rule, everyone is increased by 0.5
d <- function(data, x) data[[x]] + 0.5
psi_rule1 <- lmtp_tmle(sim_cens, a, y, time_vary = nodes,
                       cens = cens, shift = d, folds = 2)

# treatment rule, everyone is decreased by 0.5
d <- function(data, x) data[[x]] - 0.5
psi_rule2 <- lmtp_tmle(sim_cens, a, y, time_vary = nodes,
                       cens = cens, shift = d, folds = 2)

# Example 1.1
# Additive effect of rule 1 compared to a known constant
lmtp_contrast(psi_rule1, ref = 0.9)

#> Non-estimated reference value, defaulting type = 'additive'.

#> 
#>   

#> LMTP Contrast: additive

#> Null hypothesis: theta == 0

#> 
#>      theta shift ref std.error conf.low conf.high p.value
#> 1 -0.00124 0.899 0.9   0.00941  -0.0197    0.0172   0.895

# Example 1.2
# Additive effect of rule 1 compared to the population mean outcome
lmtp_contrast(psi_rule1, ref = psi_null)

#> 
#>   

#> LMTP Contrast: additive

#> Null hypothesis: theta == 0

#> 
#>   theta shift   ref std.error conf.low conf.high p.value
#> 1 0.103 0.899 0.796   0.00591   0.0914     0.115  <0.001

# Example 1.3
# Additive effects of rule 1 and 2 compared to the population mean outcome
lmtp_contrast(psi_rule1, psi_rule2, ref = psi_null)

#> 
#>   

#> LMTP Contrast: additive

#> Null hypothesis: theta == 0

#> 
#>    theta shift   ref std.error conf.low conf.high p.value
#> 1  0.103 0.899 0.796   0.00591   0.0914     0.115  <0.001
#> 2 -0.137 0.659 0.796   0.01045  -0.1577    -0.117  <0.001

# Example 1.4
# Relative risk of rule 1 compared to observed exposure
lmtp_contrast(psi_rule1, ref = psi_null, type = "rr")

#> 
#>   

#> LMTP Contrast: relative risk

#> Null hypothesis: theta == 1

#> 
#>   theta shift   ref std.error conf.low conf.high p.value
#> 1  1.13 0.899 0.796   0.00806     1.11      1.15  <0.001

# Example 1.5
# Odds of rule 1 compared to observed exposure
lmtp_contrast(psi_rule1, ref = psi_null, type = "or")

#> 
#>   

#> LMTP Contrast: odds ratio

#> Null hypothesis: theta == 1

#> 
#>   theta shift   ref std.error conf.low conf.high p.value
#> 1  2.28 0.899 0.796    0.0456     2.08      2.49  <0.001

# }