LMTP IPW Estimator

Inverse probability of treatment weighting estimator for the effects of traditional causal effects and modified treatment policies for both point treatment and longitudinal data with binary, continuous, or time-to-event outcomes. Supports binary, categorical, and continuous exposures.

lmtp_ipw(
  data,
  trt,
  outcome,
  baseline = NULL,
  time_vary = NULL,
  cens = NULL,
  k = Inf,
  id = NULL,
  shift,
  learners = NULL,
  folds = 10,
  bound = 1e-05
)

Arguments

data
A data frame in wide format containing all necessary variables for the estimation problem.

trt
A vector containing the column names of treatment variables ordered by time.

outcome
The column name of the outcome variable. In the case of time-to-event analysis, a vector containing the columns names of intermediate outcome variables and the final outcome variable ordered by time. Only numeric values are allowed. If the outcome type is binary, data should be coded as 0 and 1.

baseline
An optional vector of columns names of baseline covariates to be included for adjustment at every timepoint.

time_vary
A list the same length as the number of time points of observation with the column names for new time-varying covariates introduced at each time point. The list should be ordered following the time ordering of the model.

cens
An optional vector of column names of censoring indicators the same length as the number of time points of observation. If missingness in the outcome is present or if time-to-event outcome, must be provided.

k
An integer specifying how previous time points should be used for estimation at the given time point. Default is Inf, all time points.

id
An optional column name containing cluster level identifiers.

shift
A two argument function that specifies how treatment variables should be shifted. See examples for how to specify shift functions for continuous, binary, and categorical exposures.

learners
An optional sl3 learner stack for estimation of the exposure mechanism. If not specified, will default to using a generalized linear model.

folds
The number of folds to be used for cross-fitting. Minimum allowable number is two folds.

bound
Determines that maximum and minimum values (scaled) predictions will be bounded by. The default is 1e-5, bounding predictions by 1e-5 and 0.9999.

Value

A list of class lmtp containing the following components:

estimator

The estimator used, in this case "IPW".

theta

The estimated population LMTP effect.

standard_error

NA

low

NA

high

NA

shift

The shift function specifying the treatment policy of interest.

density_ratios

An n x Tau matrix of the estimated density ratios.

weights_r

A list the same length as folds, containing the Super Learner ensemble weights at each time-point for each fold for the propensity.

Examples

# \donttest{
library(lmtp)

# Example 1.1
# Point treatment, continuous exposure, continuous outcome, no loss-to-follow-up
# Interested in the effect of a population wide decrease in A of 5 units
# The true value under this intervention is about 519.
set.seed(56)
n <- 1000
W <- rnorm(n, 10, 5)
A <- 23 + 5*W + rnorm(n)
Y <- 7.2*A + 3*W + rnorm(n)
ex1_dat <- data.frame(W, A, Y)
d <- function(data, x) data[[x]] - 5
psi1.1 <- lmtp_ipw(ex1_dat, "A", "Y", baseline = "W", shift = d, folds = 2)
psi1.1

#> 

#> LMTP Estimator: IPW

#>    

#> Trt. Policy: (d)

#> 

#> Population intervention effect

#>       

#> Estimate: 213.3838

#>     

#> Std. error: NA

#>         

#> 95% CI: (NA, NA)

#> 

#> ! Standard errors aren't provided for the IPW estimator.

#> 

# Example 1.2
# Point treatment, continuous exposure, continuous outcome, no loss-to-follow-up
# Interested in the effect of a modified treatment policy where A is decreased by 15
# units only among observations whose observed A was above 80.
# The true value under this intervention is about 513.
d <- function(data, x) (data[[x]] > 80)*(data[[x]] - 15) + (data[[x]] <= 80)*data[[x]]
psi1.2 <- lmtp_ipw(ex1_dat, "A", "Y", baseline = "W", shift = d, folds = 2)
psi1.2

#> 

#> LMTP Estimator: IPW

#>    

#> Trt. Policy: (d)

#> 

#> Population intervention effect

#>       

#> Estimate: 271.1687

#>     

#> Std. error: NA

#>         

#> 95% CI: (NA, NA)

#> 

#> ! Standard errors aren't provided for the IPW estimator.

#> 

# Example 2.1
# Longitudinal setting, time-varying continuous exposure bounded by 0,
# time-varying covariates, and a binary outcome with no loss-to-follow-up.
# Interested in the effect of a treatment policy where exposure decreases by
# one unit at every time point if an observations observed exposure is greater
# than or equal to 2. The true value under this intervention is about 0.305.
head(sim_t4)

#>   ID L_1 A_1 L_2 A_2 L_3 A_3 L_4 A_4 Y
#> 1  1   2   3   0   1   1   1   1   3 0
#> 2  2   2   1   1   4   0   3   1   2 0
#> 3  3   1   0   1   3   1   2   1   1 1
#> 4  4   1   0   0   3   1   3   1   2 0
#> 5  5   3   3   1   1   0   1   1   2 0
#> 6  6   1   0   0   2   0   3   1   4 0

# specifying treament variables
a <- c("A_1", "A_2", "A_3", "A_4")
# specifying time varying covariates
tv <- list(c("L_1"), c("L_2"), c("L_3"), c("L_4"))
# treatment policy function to be applied at all time points
d <- function(data, trt) {
  a <- data[[trt]]
  (a - 1) * (a - 1 >= 1) + a * (a - 1 < 1)
}
progressr::with_progress({
  psi2.1 <- lmtp_ipw(sim_t4, a, "Y", time_vary = tv, shift = d, folds = 2)
})
psi2.1

#> 

#> LMTP Estimator: IPW

#>    

#> Trt. Policy: (d)

#> 

#> Population intervention effect

#>       

#> Estimate: 0.2866

#>     

#> Std. error: NA

#>         

#> 95% CI: (NA, NA)

#> 

#> ! Standard errors aren't provided for the IPW estimator.

#> 

# Example 2.2
# Example 2.1 assumed that the outcome (as well as the treatment variables)
# were directly affected by all other nodes in the past. In certain situtations,
# domain specific knowledge may suggest otherwise leading to a Markov processes.
# This can be controlled using the k argument.
progressr::with_progress({
  psi2.2 <- lmtp_ipw(sim_t4, a, "Y", time_vary = tv, shift = d,
                     k = 0, folds = 2)
})
psi2.2

#> 

#> LMTP Estimator: IPW

#>    

#> Trt. Policy: (d)

#> 

#> Population intervention effect

#>       

#> Estimate: 0.2908

#>     

#> Std. error: NA

#>         

#> 95% CI: (NA, NA)

#> 

#> ! Standard errors aren't provided for the IPW estimator.

#> 

# Example 2.3
# Using the same data as examples 2.1 and 2.2.
# Now estimating the effect of a dynamic modified treatment policy.
a <- c("A_1", "A_2", "A_3", "A_4")
time_varying <- list(c("L_1"), c("L_2"), c("L_3"), c("L_4"))
# our same shift function
shift <- function(data, trt) {
  (data[[trt]] - 1) * (data[[trt]] - 1 >= 1) + data[[trt]] * (data[[trt]] - 1 < 1)
}
# creating a dynamic mtp that applies the shift function
# but also depends on history and the current time
dynamic_mtp <- function(data, trt) {
  if (trt == "A_1") {
    # if its the first time point, follow the same mtp as before
    shift(data, trt)
  } else {
    # otherwise check if the time varying covariate equals 1
    ifelse(data[[sub("A", "L", trt)]] == 1,
           shift(data, trt), # if yes continue with the policy
           data[[trt]]) # otherwise do nothing
  }
}
psi2.3 <- lmtp_ipw(sim_t4, a, "Y", time_vary = time_varying,
                   k = 0, shift = dynamic_mtp, folds = 2)
psi2.3

#> 

#> LMTP Estimator: IPW

#>    

#> Trt. Policy: (dynamic_mtp)

#> 

#> Population intervention effect

#>       

#> Estimate: 0.2457

#>     

#> Std. error: NA

#>         

#> 95% CI: (NA, NA)

#> 

#> ! Standard errors aren't provided for the IPW estimator.

#> 

# Example 2.4
# Using the same data as examples 2.1, 2.2, and 2.3, but now treating the exposure
# as an ordered categorical variable. To account for the exposure being a
# factor we just need to modify the shift function (and the original data)
# so as to respect this.
for (i in a) {
  sim_t4[[i]] <- factor(sim_t4[[i]], levels = 0:5, ordered = TRUE)
}
d <- function(data, trt) {
  out <- list()
  a <- data[[trt]]
  for (i in 1:length(a)) {
    if (as.character(a[i]) %in% c("0", "1")) {
      out[[i]] <- as.character(a[i])
    } else {
      out[[i]] <- as.numeric(as.character(a[i])) - 1
    }
  }
  factor(unlist(out), levels = 0:5, ordered = TRUE)
}
progressr::with_progress({
  psi2.4 <- lmtp_ipw(sim_t4, a, "Y", time_vary = tv, shift = d, k = 0, folds = 2)
})
psi2.4

#> 

#> LMTP Estimator: IPW

#>    

#> Trt. Policy: (d)

#> 

#> Population intervention effect

#>       

#> Estimate: 0.3073

#>     

#> Std. error: NA

#>         

#> 95% CI: (NA, NA)

#> 

#> ! Standard errors aren't provided for the IPW estimator.

#> 

# Example 3.1
# Longitudinal setting, time-varying binary treatment, time-varying covariates
# and baseline covariates with no loss-to-follow-up. Interested in a traditional
# causal effect where treatment is set to 1 at all time points for all observations.
data("iptwExWide", package = "twang")
a <- paste0("tx", 1:3)
baseline <- c("gender", "age")
tv <- list(c("use0"), c("use1"), c("use2"))
progressr::with_progress({
  psi3.1 <-
    lmtp_ipw(iptwExWide, a, "outcome", baseline = baseline, time_vary = tv,
             shift = static_binary_on, folds = 2)
})
psi3.1

#> 

#> LMTP Estimator: IPW

#>    

#> Trt. Policy: (static_binary_on)

#> 

#> Population intervention effect

#>       

#> Estimate: -0.2969

#>     

#> Std. error: NA

#>         

#> 95% CI: (NA, NA)

#> 

#> ! Standard errors aren't provided for the IPW estimator.

#> 

# Example 4.1
# Longitudinal setting, time-varying continuous treatment, time-varying covariates,
# binary outcome with right censoring. Interested in the mean population outcome under
# the observed exposures in a hypothetical population with no loss-to-follow-up.
head(sim_cens)

#>   L1         A1 C1 L2         A2 C2 Y
#> 1  0  0.7930531  1  0  0.4098544  1 0
#> 2  1  3.9568969  1  1  7.1306144  1 1
#> 3  0 -0.9803504  1  0 -2.7858973  1 0
#> 4  1  0.8833268  1  1  1.1372421  1 1
#> 5  1  3.1666751  1  1  7.4640462  1 1
#> 6  1  3.5393771  1  1  9.5161813  1 1

a <- c("A1", "A2")
tv <- list(c("L1"), c("L2"))
cens <- c("C1", "C2")
y <- "Y"
psi4.1 <- lmtp_ipw(sim_cens, a, y, time_vary = tv, cens = cens,
                   shift = NULL, folds = 2)
psi4.1

#> 

#> LMTP Estimator: IPW

#>    

#> Trt. Policy: (NULL)

#> 

#> Population intervention effect

#>       

#> Estimate: 0.9066

#>     

#> Std. error: NA

#>         

#> 95% CI: (NA, NA)

#> 

#> ! Standard errors aren't provided for the IPW estimator.

#> 

# Example 4.2
# Using the same data as example 4.1, but now interested in the causal effect of a
# treatment policy where exposure increased by 0.5 units at all time points. The
# true value under this intervention is about 0.88.
d <- function(data, x) data[[x]] + 0.5
psi4.2 <- lmtp_ipw(sim_cens, a, y, time_vary = tv,
                   cens = cens, shift = d, folds = 2)
psi4.2

#> 

#> LMTP Estimator: IPW

#>    

#> Trt. Policy: (d)

#> 

#> Population intervention effect

#>       

#> Estimate: 1.0007

#>     

#> Std. error: NA

#>         

#> 95% CI: (NA, NA)

#> 

#> ! Standard errors aren't provided for the IPW estimator.

#> 

# Example 5.1
# Time-to-event analysis with a binary time-invariant exposure. Interested in
# the effect of treatment being given to all observations on the cumulative
# incidence of our time-to-event outcome.
a <- "trt"
# for a survival problem, the outcome arugment now takes a vector of outcomes
# if an observation experiences the event prior to the end of follow-up, all future
# outcome nodes should be set to 1 (i.e., last observation carried forward).
y <- paste0("Y.", 1:6)
cens <- paste0("C.", 0:5)
baseline <- c("W1", "W2")
progressr::with_progress({
  psi5.1 <- lmtp_ipw(sim_point_surv, a, y, baseline, cens = cens,
                     shift = static_binary_on, folds = 2)
})
psi5.1

#> 

#> LMTP Estimator: IPW

#>    

#> Trt. Policy: (static_binary_on)

#> 

#> Population intervention effect

#>       

#> Estimate: 0.4716

#>     

#> Std. error: NA

#>         

#> 95% CI: (NA, NA)

#> 

#> ! Standard errors aren't provided for the IPW estimator.

#> 

# }