LMTP Sequential Doubly Robust Estimator

LMTP Sequential Doubly Robust Estimator

lmtp_sdr(
  data,
  trt,
  outcome,
  nodes,
  baseline = NULL,
  cens = NULL,
  k = Inf,
  shift,
  outcome_type = c("binomial", "continuous"),
  bounds = NULL,
  learners_outcome = NULL,
  learners_trt = NULL,
  folds = 10,
  bound = 1e-05
)

Arguments

data
A data frame.

trt
A vector of column names of treatment variables.

outcome
The column name of the outcome variable.

nodes
A list of length tau with the column names for new nodes to be introduced at each time point. The list should be ordered following the time ordering of the model.

baseline
An optional vector of columns names for baseline covariates to be included for adjustment at every timepoint.

cens
An optional vector of column names of censoring indicators the same length as A.

k
An integer specifying how many previous time points nodes should be used for estimation at the given time point. Default is Inf, all time points.

shift
A function that specifies how tratment variables should be shifted.

outcome_type
Outcome variable type (i.e., continuous, binomial).

bounds
An optional vector of the bounds for continuous outcomes. If NULL the bounds will be taken as the minimum and maximum of the observed data. Ignored if outcome type is binary.

learners_outcome
An sl3 learner stack for estimation of the outcome regression.

learners_trt
An sl3 learner stack for estimation of the exposure mechanism.

folds
The number of folds to be used for cross-validation.

bound
Determines that maximum and minimum values (scaled) predictions will be bounded by. The default is 1e-5, bounding predictions by 1e-5 and 0.9999.

Value

A list of class lmtp containing the following components:

estimator

The estimator used, in this case "SDR".

theta

The estimated population LMTP effect.

standard_error

The estimated standard error of the LMTP effect.

low

Lower bound of the 95% confidence interval of the LMTP effect.

high

Upper bound of the 95% confidene interval of the LMTP effect.

eif

The estimated, uncentered, influence function of the estimate.

shift

weights_m

A list the same length as folds, containing the Super Learner ensemble weights at each time-point for each fold for the outcome regression.

weights_r

A list the same length as folds, containing the Super Learner ensemble weights at each time-point for each fold for the propensity.

outcome_type

The outcome variable type.

Examples

library(lmtp)

# Example 1.1
# Point treatment, continuous exposure, continuous outcome, no loss-to-follow-up
# Interested in the effect of a population wide decrease in A of 5 units
# The true value under this intervention is about 519.
set.seed(56)
n <- 1000
W <- rnorm(n, 10, 5)
A <- 23 + 5*W + rnorm(n)
Y <- 7.2*A + 3*W + rnorm(n)
ex1_dat <- data.frame(W, A, Y)
d <- function(x) x - 5
fit1.1 <- lmtp_tmle(ex1_dat, "A", "Y", list(c("W")), shift = d,
                    outcome_type = "continuous", folds = 2)

# Example 1.2
# Point treatment, continuous exposure, continuous outcome, no loss-to-follow-up
# Interesed in the effect of a modified treatment policy where A is decreased by 15
# units only among observations whose observed A was above 80.
# The true value under this intervention is about 513.
d <- function(x) (x > 80)*(x - 15) + (x <= 80)*x
fit1.2 <- lmtp_tmle(ex1_dat, "A", "Y", list(c("W")), shift = d,
                    outcome_type = "continuous", folds = 2)
fit1.2

#> 

#> LMTP Estimator: TMLE

#>    

#> Trt. Policy: (d)

#> 

#> Population intervention effect

#>       

#> Estimate: 507.768

#>     

#> Std. error: 4.8914

#>         

#> 95% CI: (498.1811, 517.3549)

#> 

# Example 2.1
# Longitudinal setting, time-varying continuous exposure bounded by 0,
# time-varying covariates, and a binary outcome with no loss-to-follow-up.
# Interested in the effect of a treatment policy where exposure decreases by
# one unit at every time point if an observations observed exposure is greater
# than or equal to 2. The true value under this intervention is about 0.305.
head(sim_t4)

#>   ID L_1 A_1 L_2 A_2 L_3 A_3 L_4 A_4 Y
#> 1  1   2   3   0   1   1   1   1   3 0
#> 2  2   2   1   1   4   0   3   1   2 0
#> 3  3   1   0   1   3   1   2   1   1 1
#> 4  4   1   0   0   3   1   3   1   2 0
#> 5  5   3   3   1   1   0   1   1   2 0
#> 6  6   1   0   0   2   0   3   1   4 0

# specifying treament variables
a <- c("A_1", "A_2", "A_3", "A_4")
# specifying time varying covariates
time_varying <- list(c("L_1"), c("L_2"), c("L_3"), c("L_4"))
# treatment policy function to be applied at all time points
d <- function(a) {
  (a - 1) * (a - 1 >= 1) + a * (a - 1 < 1)
}
progressr::with_progress({
  fit2.1 <- lmtp_tmle(sim_t4, a, "Y", time_varying, shift = d, folds = 2)
})
fit2.1

#> 

#> LMTP Estimator: TMLE

#>    

#> Trt. Policy: (d)

#> 

#> Population intervention effect

#>       

#> Estimate: 0.2644

#>     

#> Std. error: 0.017

#>         

#> 95% CI: (0.2311, 0.2976)

#> 

# Example 2.2
# Example 2.1 assumed that the outcome (as well as the treatment variables)
# were directly affected by all other nodes in the past. In certain situtations,
# domain specific knowledge may suggest otherwise leading to a Markov processes.
# This can be controlled using the k argument.
progressr::with_progress({
  fit2.2 <- lmtp_tmle(sim_t4, a, "Y", time_varying, shift = d,
                      k = 1, folds = 2)
})
fit2.2

#> 

#> LMTP Estimator: TMLE

#>    

#> Trt. Policy: (d)

#> 

#> Population intervention effect

#>       

#> Estimate: 0.2632

#>     

#> Std. error: 0.0168

#>         

#> 95% CI: (0.2303, 0.2962)

#> 

# Example 3.1
# Longitudinal setting, time-varying binary treatment, time-varying covariates
# and baseline covariates with no loss-to-follow-up. Interested in a traditional
# causal effect where treatment is set to 1 at all time points for all observations.
data("iptwExWide", package = "twang")
a <- paste0("tx", 1:3)
baseline <- c("gender", "age")
nodes <- list(c("use0"), c("use1"), c("use2"))
progressr::with_progress({
  fit3.1 <-
    lmtp_tmle(iptwExWide, a, "outcome", nodes, baseline = baseline,
              shift = function(x) 1, outcome_type = "continuous",
              folds = 2)
})
fit3.1

#> 

#> LMTP Estimator: TMLE

#>    

#> Trt. Policy: (function(x) 1)

#> 

#> Population intervention effect

#>       

#> Estimate: -0.2866

#>     

#> Std. error: 0.0419

#>         

#> 95% CI: (-0.3687, -0.2046)

#> 

# Example 4.1
# Longitudinal setting, time-varying continuous treatment, time-varying covariates,
# binary outcome with right censoring. Interested in the mean population outcome under
# the observed exposures in a hypothetical population with no loss-to-follow-up.
head(sim_cens)

#>   L1         A1 C1 L2         A2 C2 Y
#> 1  0  0.7930531  1  0  0.4098544  1 0
#> 2  1  3.9568969  1  1  7.1306144  1 1
#> 3  0 -0.9803504  1  0 -2.7858973  1 0
#> 4  1  0.8833268  1  1  1.1372421  1 1
#> 5  1  3.1666751  1  1  7.4640462  1 1
#> 6  1  3.5393771  1  1  9.5161813  1 1

a <- c("A1", "A2")
nodes <- list(c("L1"), c("L2"))
cens <- c("C1", "C2")
y <- "Y"
fit4.1 <- lmtp_tmle(sim_cens, a, y, nodes, cens = cens, shift = NULL, folds = 2)
fit4.1

#> 

#> LMTP Estimator: TMLE

#>    

#> Trt. Policy: (NULL)

#> 

#> Population intervention effect

#>       

#> Estimate: 0.7964

#>     

#> Std. error: 0.0129

#>         

#> 95% CI: (0.771, 0.8217)

#> 

# Example 4.2
# Using the same data as example 4.1, but now interested in the causal effect of a
# modified treatment policy where ...

# Example 5.1
# Time-to-event analysis with a time-invariant exposure.

# Example 5.2
# Time-to-event analysis with a time-varying exposure and time-varying covariates.