LMTP Targeted Maximum Likelihood Estimator

LMTP Targeted Maximum Likelihood Estimator

lmtp_tmle(
  data,
  trt,
  outcome,
  nodes,
  baseline = NULL,
  cens = NULL,
  k = Inf,
  shift,
  outcome_type = c("binomial", "continuous"),
  bounds = NULL,
  learners_outcome = NULL,
  learners_trt = NULL,
  progress_bar = TRUE
)

Arguments

data
A data frame.

trt
A vector of column names for treatment variables.

outcome
The column name of the outcome variable.

nodes
A list of length tau with the column names for new nodes to be introduced at each time point. The list should be ordered following the time ordering of the model.

baseline
An optional vector of columns names for baseline covariates to be included for adjustment at every timepoint. If k = Inf, should be NULL and these variables should be added to the first index of nodes.

cens
An optional vector of column names of censoring indicators the same length as A.

k
An integer specifying how many previous time points nodes should be used for estimation at the given time point. Default is Inf, all time points.

shift
A function that specifies how tratment variables should be shifted.

outcome_type
Outcome variable type (i.e., continuous, binomial).

bounds
An optional vector of the bounds for continuous outcomes. If NULL the bounds will be taken as the minimum and maximum of the observed data. Ignored if outcome type is binary.

learners_outcome
An sl3 learner stack for estimation of the outcome regression.

learners_trt
An sl3 learner stack for estimation of the exposure mechanism.

progress_bar
Should a progress bar be displayed? Default is TRUE.

Value

TODO

Examples

# TO DO